refactor: reorganize rag modules
