+++ docs/rag-system-design.md
@@ -64,7 +64,7 @@
 ## 6. Go 代码模块划分
 
 ```
-server/rag/
+internal/rag/
 ├── config/          # 仓库与模型配置
 ├── sync/            # GitHub 同步逻辑
 ├── ingest/          # Markdown 解析与分块
