+++ internal/rag/config/runtime.go
