+++ internal/rag/embed/bge.go
