+++ internal/rag/embed/openai.go
