+++ internal/rag/ingest/hash.go
