+++ internal/rag/ingest/walk.go
